+++ app/ui/src/util/http/settings.jsx
@@ -15,6 +15,8 @@ export const fetchSettings = async ({ signal }) => {
 export const updateSettings = async ({ settings }) => {
   const url = `${API_URL}/settings/general`;
   settings.licenses_expiration.warning_period = parseInt(settings.licenses_expiration.warning_period) || 90;
+  settings.previous_experience.landings_day = parseInt(settings.previous_experience.landings_day) || 0;
+  settings.previous_experience.landings_night = parseInt(settings.previous_experience.landings_night) || 0;
   const options = {
     method: 'PUT',
     headers: { 'Authorization': `Bearer ${getAuthToken()}`, 'Content-Type': 'application/json' },