+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [4.6.0] - 26.06.2026
 
 - Update: Golang and npm packages/security patches.
 - New: Added support for tracking previous flight experience on the Settings page. Entered flight experience is automatically included in Logbook totals, Dashboard statistics, and PDF exports.